add metadata for help and basic methods implementation
This commits extends pyang-swagger in order to add some metadata in the
rest description that is then used by swagger-code gen to implement the
help and to provide a basic implementation for some common functions.

For the help, pyang adds the "x-is-key" to indicate if a leaf is the key
of a container of elements

For the basic functions implementation, it adds the "x-has-default-impl"
flag to the functions that should have a default implementation in the
code automatically generated.